copy assets on non-development device

+++ proxyChecker.php
@@ -119,17 +119,19 @@ setFilePermissions([$untestedFilePath, $workingPath, $deadPath]);
 
 // move backup added proxies
 
-$assets = getFilesByExtension(__DIR__ . '/assets/proxies', 'txt');
-$assets[] = __DIR__ . '/proxies-backup.txt';
-foreach ($assets as $asset) {
-  echo $asset . ' - ' . (file_exists($asset) ? 'true' : 'false') . PHP_EOL;
-  if (file_exists($asset)) {
-    $move = moveContent($asset, $untestedFilePath);
-    if ($move === '') {
-      unlink($asset);
-      exit('copied assets');
-    } else {
-      echo $move . PHP_EOL;
+if (gethostname() !== 'DESKTOP-JVTSJ6I') {
+  $assets = getFilesByExtension(__DIR__ . '/assets/proxies', 'txt');
+  $assets[] = __DIR__ . '/proxies-backup.txt';
+  foreach ($assets as $asset) {
+    echo $asset . ' - ' . (file_exists($asset) ? 'true' : 'false') . PHP_EOL;
+    if (file_exists($asset)) {
+      $move = moveContent($asset, $untestedFilePath);
+      if ($move === '') {
+        unlink($asset);
+        exit('copied assets');
+      } else {
+        echo $move . PHP_EOL;
+      }
     }
   }
 }
